requirements.txt
# Dependencies
fastapi
uvicorn
litellm
python-dotenv
chromadb
streamlit
pandas
